+++ mtc-model/src/i18n/en_US.rs
@@ -46,7 +46,6 @@ pub static EN_US: &str = r#"{
             "name": "name",
             "new_field": "new field",
             "no": "No",
-            "oops": "Oops!",
             "page": "page",
             "password": "password",
             "password_change": "Change password",
@@ -55,7 +54,6 @@ pub static EN_US: &str = r#"{
             "password_new": "new password",
             "password_old": "old password",            
             "permissions": "Permissions",
-            "personas": "Personas",
             "processing": "Processing",
             "published": "Published",
             "rank": "rank",
@@ -71,15 +69,11 @@ pub static EN_US: &str = r#"{
             "single": "single",
             "singles": "Single",
             "slug": "slug",
-            "something_wrong": "Something went wrong",
             "title": "title",
-            "try_again": "Try again",
             "type": "type",
             "updated_at": "updated at",
             "upload": "upload",
-            "users": "Users",
-            "user_active": "Active",
-            "user_blocked": "Blocked",          
+            "users": "Users",            
             "welcome": "Welcome!",
             "yes": "Yes"
         },
