+++ mtc-model/src/i18n/uk_UA.rs
@@ -46,7 +46,6 @@ pub static UK_UA: &str = r#"{
             "name": "ПІБ",
             "new_field": "нове поле",
             "no": "Ні",
-            "oops": "Упс!",
             "page": "сторінка",
             "password": "пароль",
             "password_change": "Змінити пароль",
@@ -55,7 +54,6 @@ pub static UK_UA: &str = r#"{
             "password_new": "новий пароль",
             "password_old": "дійсний пароль", 
             "permissions": "Дозволи",
-            "personas": "Особовий склад",
             "processing": "Опрацювання",
             "published": "Опубліковано",
             "rank": "звання",
@@ -71,15 +69,11 @@ pub static UK_UA: &str = r#"{
             "single": "одиночна",
             "singles": "Одиночні",
             "slug": "ідентифікатор",
-            "something_wrong": "Невдача спіткала нашу команду",
             "title": "назва",
-            "try_again": "Спробувати ще",
             "type": "тип",
             "upload": "завантажити",
             "updated_at": "оновлено",
             "users": "Користувачі",
-            "user_active": "Чинний",
-            "user_blocked": "Заблокований",
             "welcome": "Вітаємо!",
             "yes": "Так"
         },
