+++ mtc-web/src/element/editor/html_field.rs
@@ -1,5 +1,5 @@
 use dioxus::prelude::*;
-use crate::page::administrator::editor::FieldProps;
+use crate::element::editor::FieldProps;
 
 #[component]
 pub fn HtmlField(props: FieldProps) -> Element {
