+++ mtc-web/src/element/mod.rs
@@ -2,3 +2,4 @@ pub mod user_dashboard;
 pub mod header;
 pub mod footer;
 pub mod migration;
+pub mod editor;
